+++ read.c
@@ -15,7 +15,7 @@
  */
 
 #include <string.h>
-#include <stdio.h>
+#include <stdint.h>
 
 static int
 readname_loop(char *packet, int packetlen, char **src, char *dst, size_t length, size_t loop)
@@ -95,19 +95,20 @@ readshort(char *packet, char **src, short *dst)
 }
 
 int
-readlong(char *packet, char **src, long *dst)
+readlong(char *packet, char **src, uint32_t *dst)
 {
+	// A long as described in dns protocol is always 32 bits
 	unsigned char *p;
 
 	p = *src;
 
-	*dst = ((long)p[0] << 24) 
-		 | ((long)p[1] << 16) 
-		 | ((long)p[2] << 8)
-		 | ((long)p[3]);
+	*dst = ((uint32_t)p[0] << 24) 
+		 | ((uint32_t)p[1] << 16) 
+		 | ((uint32_t)p[2] << 8)
+		 | ((uint32_t)p[3]);
 
-	(*src) += sizeof(long);
-	return sizeof(long);
+	(*src) += sizeof(uint32_t);
+	return sizeof(uint32_t);
 }
 
 int
@@ -147,8 +148,9 @@ putshort(char **dst, short value)
 }
 
 int
-putlong(char **dst, long value)
+putlong(char **dst, uint32_t value)
 {
+	// A long as described in dns protocol is always 32 bits
 	unsigned char *p;
 
 	p = *dst;
@@ -159,7 +161,7 @@ putlong(char **dst, long value)
 	*p++ = (value);
 
 	(*dst) = p;
-	return sizeof(long);
+	return sizeof(uint32_t);
 }
 
 int
